margin_based_water_marking/requirements.txt
Package                  Version
------------------------ -----------
blessed                  1.20.0
certifi                  2024.8.30
charset-normalizer       3.4.0
click                    8.1.7
contourpy                1.3.0
cycler                   0.12.1
docker-pycreds           0.4.0
filelock                 3.16.1
fonttools                4.54.1
fsspec                   2024.9.0
gitdb                    4.0.11
GitPython                3.1.43
gpustat                  1.1.1
idna                     3.10
Jinja2                   3.1.4
joblib                   1.4.2
kiwisolver               1.4.7
MarkupSafe               3.0.1
matplotlib               3.9.2
mpmath                   1.3.0
networkx                 3.4.1
numpy                    2.1.2
nvidia-cublas-cu12       12.1.3.1
nvidia-cuda-cupti-cu12   12.1.105
nvidia-cuda-nvrtc-cu12   12.1.105
nvidia-cuda-runtime-cu12 12.1.105
nvidia-cudnn-cu12        9.1.0.70
nvidia-cufft-cu12        11.0.2.54
nvidia-curand-cu12       10.3.2.106
nvidia-cusolver-cu12     11.4.5.107
nvidia-cusparse-cu12     12.1.0.106
nvidia-ml-py             12.560.30
nvidia-nccl-cu12         2.20.5
nvidia-nvjitlink-cu12    12.6.77
nvidia-nvtx-cu12         12.1.105
packaging                24.1
pillow                   11.0.0
pip                      24.2
platformdirs             4.3.6
plotly                   5.24.1
protobuf                 5.28.2
psutil                   6.0.0
pyparsing                3.2.0
python-dateutil          2.9.0.post0
PyYAML                   6.0.2
requests                 2.32.3
scikit-learn             1.5.2
scipy                    1.14.1
sentry-sdk               2.17.0
setproctitle             1.3.3
setuptools               75.1.0
six                      1.16.0
smmap                    5.0.1
sympy                    1.13.3
tenacity                 9.0.0
threadpoolctl            3.5.0
torch                    2.4.1
torchvision              0.19.1
triton                   3.0.0
typing_extensions        4.12.2
urllib3                  2.2.3
waitGPU                  0.0.3
wandb                    0.18.5
wcwidth                  0.2.13